chore: move api connector file
